refactor: config, exception, util을 global 패키지로 이동
