+++ src/main/java/roomescape/global/config/WebConfig.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
