+++ src/main/java/roomescape/global/exception/ApplicationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class ApplicationException extends RuntimeException {
