+++ src/main/java/roomescape/global/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
