+++ src/main/java/roomescape/time/exception/TimeErrorCode.java
@@ -1,7 +1,7 @@
 package roomescape.time.exception;
 
 import org.springframework.http.HttpStatus;
-import roomescape.exception.ErrorCode;
+import roomescape.global.exception.ErrorCode;
 
 public enum TimeErrorCode implements ErrorCode {
